spring 포팅 dev로 배포

+++ frontend/src/api/axios/axios.custom.ts
@@ -2,7 +2,7 @@ import CabinetStatus from "@/types/enum/cabinet.status.enum";
 import CabinetType from "@/types/enum/cabinet.type.enum";
 import instance from "@/api/axios/axios.instance";
 
-const axiosLogoutUrl = "/auth/logout";
+const axiosLogoutUrl = "/v4/auth/logout";
 export const axiosLogout = async (): Promise<any> => {
   try {
     const response = await instance.get(axiosLogoutUrl);
@@ -12,7 +12,7 @@ export const axiosLogout = async (): Promise<any> => {
   }
 };
 
-const axiosMyInfoURL = "/api/my_info";
+const axiosMyInfoURL = "/v4/users/me";
 export const axiosMyInfo = async (): Promise<any> => {
   try {
     const response = await instance.get(axiosMyInfoURL);
@@ -21,8 +21,8 @@ export const axiosMyInfo = async (): Promise<any> => {
     throw error;
   }
 };
-
-const axiosUpdateCabinetMemoURL = "/api/lent/update_cabinet_memo";
+/*
+const axiosUpdateCabinetMemoURL = "/v4/lent/me/memo";
 export const axiosUpdateCabinetMemo = async (
   cabinet_memo: object
 ): Promise<any> => {
@@ -37,40 +37,39 @@ export const axiosUpdateCabinetMemo = async (
   }
 };
 
-const axiosUpdateCabinetTitleURL = "/api/lent/update_cabinet_title";
+const axiosUpdateCabinetTitleURL = "/v4/lent/me/cabinet-title";
 export const axiosUpdateCabinetTitle = async (
-  cabinet_title: object
+  cabinet_title: any
 ): Promise<any> => {
   try {
-    const response = await instance.patch(
-      axiosUpdateCabinetTitleURL,
-      cabinet_title
-    );
+    const response = await instance.patch(axiosUpdateCabinetTitleURL, {
+      cabinetTitle: cabinet_title.title,
+    });
     return response;
   } catch (error) {
     throw error;
   }
-};
+}; */
 
 // V3 API
-const axiosLocationFloorURL = "/api/cabinet_info";
-export const axiosLocationFloor = async (): Promise<any> => {
+const axiosBuildingFloorURL = "/v4/cabinets/buildings/floors";
+export const axiosBuildingFloor = async (): Promise<any> => {
   try {
-    const response = await instance.get(axiosLocationFloorURL);
+    const response = await instance.get(axiosBuildingFloorURL);
     return response;
   } catch (error) {
     throw error;
   }
 };
 
-const axiosCabinetByLocationFloorURL = "/api/cabinet_info/";
-export const axiosCabinetByLocationFloor = async (
-  location: string,
+const axiosCabinetByBuildingFloorURL = "/v4/cabinets/buildings/";
+export const axiosCabinetByBuildingFloor = async (
+  Building: string,
   floor: number
 ): Promise<any> => {
   try {
     const response = await instance.get(
-      `${axiosCabinetByLocationFloorURL}${location}/${floor}`
+      `${axiosCabinetByBuildingFloorURL}${Building}/floors/${floor}`
     );
     return response;
   } catch (error) {
@@ -78,8 +77,11 @@ export const axiosCabinetByLocationFloor = async (
   }
 };
 
-const axiosCabinetByIdURL = "/api/cabinet_info/";
-export const axiosCabinetById = async (cabinetId: number): Promise<any> => {
+const axiosCabinetByIdURL = "/v4/cabinets/";
+export const axiosCabinetById = async (
+  cabinetId: number | null
+): Promise<any> => {
+  if (cabinetId === null) return;
   try {
     const response = await instance.get(`${axiosCabinetByIdURL}${cabinetId}`);
     return response;
@@ -88,8 +90,9 @@ export const axiosCabinetById = async (cabinetId: number): Promise<any> => {
   }
 };
 
-const axiosLentIdURL = "/api/lent/";
-export const axiosLentId = async (cabinetId: number): Promise<any> => {
+const axiosLentIdURL = "/v4/lent/cabinets/";
+export const axiosLentId = async (cabinetId: number | null): Promise<any> => {
+  if (cabinetId === null) return;
   try {
     const response = await instance.post(`${axiosLentIdURL}${cabinetId}`);
     return response;
@@ -98,7 +101,7 @@ export const axiosLentId = async (cabinetId: number): Promise<any> => {
   }
 };
 
-const axiosMyLentInfoURL = "/api/my_lent_info";
+const axiosMyLentInfoURL = "/v4/lent/me";
 export const axiosMyLentInfo = async (): Promise<any> => {
   try {
     const response = await instance.get(axiosMyLentInfoURL);
@@ -108,32 +111,48 @@ export const axiosMyLentInfo = async (): Promise<any> => {
   }
 };
 
-const axiosReturnURL = "/api/lent/return";
-export const axiosReturn = async (): Promise<any> => {
+const axiosUpdateMyCabinetInfoURL = "/v4/lent/me/cabinet";
+export const axiosUpdateMyCabinetInfo = async (
+  title: string | null,
+  memo: string | null
+): Promise<any> => {
   try {
-    const response = await instance.delete(axiosReturnURL);
+    const response = await instance.patch(axiosUpdateMyCabinetInfoURL, {
+      title,
+      memo,
+    });
     return response;
   } catch (error) {
     throw error;
   }
 };
 
-const axiosSendCabinetPasswordURL = "/api/lent/return-memo";
+const axiosReturnURL = "/v4/lent/return";
+export const axiosReturn = async (): Promise<any> => {
+  try {
+    const response = await instance.patch(axiosReturnURL);
+    return response;
+  } catch (error) {
+    throw error;
+  }
+};
+
+const axiosSendCabinetPasswordURL = "/v4/lent/return-memo";
 export const axiosSendCabinetPassword = async (password: string) => {
   try {
     const response = await instance.patch(axiosSendCabinetPasswordURL, {
-      cabinet_memo: password,
+      cabinetMemo: password,
     });
   } catch (error) {
     throw error;
   }
 };
 
-const axiosMyLentLogURL = "/api/my_lent_info/log";
+const axiosMyLentLogURL = "/v4/lent/me/histories";
 export const axiosMyLentLog = async (page: number): Promise<any> => {
   try {
     const response = await instance.get(axiosMyLentLogURL, {
-      params: { page: page, length: 10 },
+      params: { page: page, size: 10 },
     });
     return response;
   } catch (error) {
@@ -142,7 +161,7 @@ export const axiosMyLentLog = async (page: number): Promise<any> => {
 };
 
 // Admin API
-const axiosAdminAuthLoginURL = "/api/admin/auth/login";
+const axiosAdminAuthLoginURL = "/v4/admin/auth/login";
 export const axiosAdminAuthLogin = async (
   id: string,
   password: string
@@ -158,10 +177,11 @@ export const axiosAdminAuthLogin = async (
   }
 };
 
-const axiosAdminCabinetInfoByIdURL = "/api/admin/cabinet/";
+const axiosAdminCabinetInfoByIdURL = "/v4/admin/cabinets/";
 export const axiosAdminCabinetInfoByCabinetId = async (
-  cabinetId: number
+  cabinetId: number | null
 ): Promise<any> => {
+  if (cabinetId === null) return;
   try {
     const response = await instance.get(
       axiosAdminCabinetInfoByIdURL + cabinetId
@@ -172,66 +192,78 @@ export const axiosAdminCabinetInfoByCabinetId = async (
   }
 };
 
-const axiosGetCabinetStateURL = "/api/admin/cabinet/count/floor";
-export const axiosGetCabinetState = async (): Promise<any> => {
-  try {
-    const response = await instance.get(axiosGetCabinetStateURL);
-    return response;
-  } catch (error) {
-    throw error;
-  }
-};
-
-const axiosAdminReturnURL = "/api/admin/return/cabinet/";
+/*
+const axiosAdminReturnURL = "/v4/admin/return-cabinets/";
 export const axiosAdminReturn = async (cabinetId: number): Promise<any> => {
   try {
-    const response = await instance.delete(axiosAdminReturnURL + cabinetId);
+    const response = await instance.patch(axiosAdminReturnURL + cabinetId);
     return response;
   } catch (error) {
     throw error;
   }
 };
-
-const axiosReturnByUserIdURL = "/api/admin/return/user/";
+ */
+const axiosReturnByUserIdURL = "/v4/admin/return-users/";
 export const axiosReturnByUserId = async (userId: number): Promise<any> => {
   try {
-    const response = await instance.delete(axiosReturnByUserIdURL + userId);
+    const response = await instance.patch(axiosReturnByUserIdURL + userId);
     return response;
   } catch (error) {
     throw error;
   }
 };
 
-const axiosBundleReturnURL = "/api/admin/return/bundle/cabinet";
+const axiosBundleReturnURL = "/v4/admin/return-cabinets";
 export const axiosBundleReturn = async (
   cabinetIdList: number[]
 ): Promise<any> => {
   try {
-    const response = await instance.delete(axiosBundleReturnURL, {
-      data: cabinetIdList,
+    const response = await instance.patch(axiosBundleReturnURL, {
+      cabinetIds: cabinetIdList,
     });
     return response;
   } catch (error) {
     throw error;
   }
 };
-
-const axiosUpdateCabinetTypeURL = "/api/admin/cabinet/lentType/";
+/*
+const axiosUpdateCabinetTypeURL = "/v4/admin/cabinets/lent-types/";
 export const axiosUpdateCabinetType = async (
   cabinetId: number,
   lentType: CabinetType
 ) => {
   try {
     const response = await instance.patch(
-      `${axiosUpdateCabinetTypeURL}${cabinetId}/${lentType}`
+      `${axiosUpdateCabinetTypeURL}${lentType}`,
+      {
+        cabinetIds: [cabinetId],
+      }
     );
     return response;
   } catch (error) {
     throw error;
   }
-};
+}; */
 
-const axiosBundleUpdateCabinetTypeURL = "/api/admin/cabinet/bundle/lentType/";
+const axiosUpdateCabinetsURL = "/v4/admin/cabinets/";
+export const axiosUpdateCabinets = async (
+  cabinetIds: number[],
+  lentType: CabinetType | null,
+  status: CabinetStatus | null
+): Promise<any> => {
+  try {
+    const response = await instance.patch(axiosUpdateCabinetsURL, {
+      cabinetIds,
+      lentType,
+      status,
+    });
+    return response;
+  } catch (error) {
+    throw error;
+  }
+};
+/*
+const axiosBundleUpdateCabinetTypeURL = "/v4/admin/cabinets/lent-types/";
 export const axiosBundleUpdateCabinetType = async (
   cabinetIdList: number[],
   cabinetType: CabinetType
@@ -239,82 +271,8 @@ export const axiosBundleUpdateCabinetType = async (
   try {
     const response = await instance.patch(
       `${axiosBundleUpdateCabinetTypeURL}${cabinetType}`,
-      cabinetIdList
-    );
-    return response;
-  } catch (error) {
-    throw error;
-  }
-};
-
-const axiosUpdateCabinetStatusURL = "/api/admin/cabinet/status/";
-export const axiosUpdateCabinetStatus = async (
-  cabinetId: number,
-  status: CabinetStatus
-): Promise<any> => {
-  try {
-    const response = await instance.patch(
-      `${axiosUpdateCabinetStatusURL}${cabinetId}/${status}`
-    );
-    return response;
-  } catch (error) {
-    throw error;
-  }
-};
-
-const axiosBundleUpdateCabinetStatusURL = "/api/admin/cabinet/bundle/status/";
-export const axiosBundleUpdateCabinetStatus = async (
-  cabinetIdList: number[],
-  cabinetStatus: CabinetStatus
-): Promise<any> => {
-  try {
-    const response = await instance.patch(
-      `${axiosBundleUpdateCabinetStatusURL}${cabinetStatus}`,
-      cabinetIdList
-    );
-    return response;
-  } catch (error) {
-    throw error;
-  }
-};
-
-const axiosSearchByIntraIdURL = "/api/admin/search/intraId/";
-export const axiosSearchByIntraId = async (intraId: string) => {
-  try {
-    const response = await instance.get(axiosSearchByIntraIdURL + intraId, {
-      params: { page: 0, length: 10 },
-    });
-    return response;
-  } catch (error) {
-    throw error;
-  }
-};
-
-const axiosSearchByCabinetNumURL = "/api/admin/search/cabinet/visibleNum/";
-export const axiosSearchByCabinetNum = async (
-  number: number,
-  floor?: number
-) => {
-  try {
-    const response = await instance.get(axiosSearchByCabinetNumURL + number, {
-      params: { floor: floor },
-    });
-    return response;
-  } catch (error) {
-    throw error;
-  }
-};
-
-const axiosSearchDetailByIntraIdURL = "/api/admin/search/";
-export const axiosSearchDetailByIntraId = async (
-  intraId: string,
-  page: number
-) => {
-  try {
-    const response = await instance.get(
-      axiosSearchDetailByIntraIdURL + intraId,
       {
-        params: { page: page, length: 10 },
+        cabinetIds: cabinetIdList,
       }
     );
     return response;
@@ -322,12 +280,89 @@ export const axiosSearchDetailByIntraId = async (
     throw error;
   }
 };
+ */
+/*
+const axiosUpdateCabinetStatusURL = "/v4/admin/cabinets/status/";
+export const axiosUpdateCabinetStatus = async (
+  cabinetId: number,
+  status: CabinetStatus
+): Promise<any> => {
+  try {
+    const response = await instance.patch(
+      `${axiosUpdateCabinetStatusURL}${status}`,
+      {
+        cabinetIds: [cabinetId],
+      }
+    );
+    return response;
+  } catch (error) {
+    throw error;
+  }
+}; */
+/*
+const axiosBundleUpdateCabinetStatusURL = "/v4/admin/cabinets/status/";
+export const axiosBundleUpdateCabinetStatus = async (
+  cabinetIdList: number[],
+  cabinetStatus: CabinetStatus
+): Promise<any> => {
+  try {
+    const response = await instance.patch(
+      `${axiosBundleUpdateCabinetStatusURL}${cabinetStatus}`,
+      {
+        cabinetIds: cabinetIdList,
+      }
+    );
+    return response;
+  } catch (error) {
+    throw error;
+  }
+}; */
 
-const axiosDeleteCurrentBanLogURL = "/api/admin/log/ban/";
-export const axiosDeleteCurrentBanLog = async (userId: number) => {
+const axiosSearchByIntraIdURL = "/v4/admin/search/users-simple";
+export const axiosSearchByIntraId = async (intraId: string) => {
+  try {
+    const response = await instance.get(axiosSearchByIntraIdURL, {
+      params: { name: intraId, page: 0, size: 10 },
+    });
+    return response;
+  } catch (error) {
+    throw error;
+  }
+};
+
+const axiosSearchByCabinetNumURL = "/v4/admin/search/cabinets";
+export const axiosSearchByCabinetNum = async (number: number) => {
+  try {
+    const response = await instance.get(axiosSearchByCabinetNumURL, {
+      params: { visibleNum: number },
+    });
+    return response;
+  } catch (error) {
+    throw error;
+  }
+};
+
+const axiosSearchDetailByIntraIdURL = "/v4/admin/search/users";
+export const axiosSearchDetailByIntraId = async (
+  intraId: string,
+  page: number
+) => {
+  try {
+    const response = await instance.get(axiosSearchDetailByIntraIdURL, {
+      params: { name: intraId, page: page, size: 10 },
+    });
+    return response;
+  } catch (error) {
+    throw error;
+  }
+};
+
+const axiosDeleteCurrentBanLogURL = "/v4/admin/users/";
+export const axiosDeleteCurrentBanLog = async (userId: number | null) => {
+  if (userId === null) return;
   try {
     const response = await instance.delete(
-      axiosDeleteCurrentBanLogURL + userId.toString()
+      axiosDeleteCurrentBanLogURL + userId?.toString() + "/ban-history"
     );
     return response;
   } catch (error) {
@@ -335,11 +370,11 @@ export const axiosDeleteCurrentBanLog = async (userId: number) => {
   }
 };
 
-const axiosGetBrokenCabinetListURL = "/api/admin/search/cabinet/broken/";
+const axiosGetBrokenCabinetListURL = "/v4/admin/cabinets/status/BROKEN";
 export const axiosGetBrokenCabinetList = async () => {
   try {
     const response = await instance.get(axiosGetBrokenCabinetListURL, {
-      params: { page: 0, length: 0 },
+      params: { page: 0, size: 0 },
     });
     return response.data.result;
   } catch (error) {
@@ -347,11 +382,11 @@ export const axiosGetBrokenCabinetList = async () => {
   }
 };
 
-const axiosGetBannedUserListURL = "/api/admin/search/user/banned/";
+const axiosGetBannedUserListURL = "v4/admin/statistics/users/banned";
 export const axiosGetBannedUserList = async () => {
   try {
     const response = await instance.get(axiosGetBannedUserListURL, {
-      params: { page: 0, length: 0 },
+      params: { page: 0, size: 0 },
     });
     return response.data.result;
   } catch (error) {
@@ -359,11 +394,11 @@ export const axiosGetBannedUserList = async () => {
   }
 };
 
-const axiosGetStatisticsURL = "/api/admin/search/cabinet/statistics/";
+const axiosGetStatisticsURL = "/v4/admin/statistics/lent-histories";
 export const axiosGetStatistics = async (start: number, end: number) => {
   try {
     const response = await instance.get(axiosGetStatisticsURL, {
-      params: { start: start, end: end },
+      params: { startDate: start, endDate: end },
     });
     return response.data;
   } catch (error) {
@@ -371,7 +406,8 @@ export const axiosGetStatistics = async (start: number, end: number) => {
   }
 };
 
-const axiosGetCabinetNumbersPerFloorURL = "/api/admin/cabinet/count/floor";
+const axiosGetCabinetNumbersPerFloorURL =
+  "/v4/admin/statistics/buildings/floors/cabinets";
 export const axiosGetCabinetNumbersPerFloor = async () => {
   try {
     const response = await instance.get(axiosGetCabinetNumbersPerFloorURL);
@@ -381,11 +417,11 @@ export const axiosGetCabinetNumbersPerFloor = async () => {
   }
 };
 
-const axiosGetOverdueUserListURL = "api/admin/search/user/overdue/";
+const axiosGetOverdueUserListURL = "/v4/admin/statistics/users/overdue";
 export const axiosGetOverdueUserList = async () => {
   try {
     const response = await instance.get(axiosGetOverdueUserListURL, {
-      params: { page: 0, length: 0 },
+      params: { page: 0, size: 0 },
     });
     return response.data.result;
   } catch (error) {
@@ -393,16 +429,17 @@ export const axiosGetOverdueUserList = async () => {
   }
 };
 
-const axiosGetCabinetLentLogURL = "/api/admin/log/cabinet/";
+const axiosGetCabinetLentLogURL = "/v4/admin/cabinets/";
 export const axiosGetCabinetLentLog = async (
-  cabinetId: number,
+  cabinetId: number | null,
   page: number
 ): Promise<any> => {
+  if (cabinetId == null) return;
   try {
     const response = await instance.get(
-      axiosGetCabinetLentLogURL + cabinetId.toString(),
+      axiosGetCabinetLentLogURL + cabinetId.toString() + "/lent-histories",
       {
-        params: { page: page, length: 10 },
+        params: { page: page, size: 10 },
       }
     );
     return response;
@@ -411,16 +448,17 @@ export const axiosGetCabinetLentLog = async (
   }
 };
 
-const axiosGetUserLentLogURL = "/api/admin/log/user/";
+const axiosGetUserLentLogURL = "/v4/admin/users/";
 export const axiosGetUserLentLog = async (
-  userId: number,
+  userId: number | null,
   page: number
 ): Promise<any> => {
+  if (userId == null) return;
   try {
     const response = await instance.get(
-      axiosGetUserLentLogURL + userId.toString(),
+      axiosGetUserLentLogURL + userId.toString() + "/lent-histories",
       {
-        params: { page: page, length: 10 },
+        params: { page: page, size: 10 },
       }
     );
     return response;
